services/ai-proxy-service/requirements.txt
fastapi==0.121.3
uvicorn[standard]==0.30.6
pydantic==2.10.6
pydantic-settings==2.7.0
google-generativeai==0.8.3
httpx==0.28.1
python-dotenv==1.0.1
